refactor(lazer): restructure protocol modules
