+++ lazer/sdk/rust/protocol/src/jrpc.rs
@@ -1,8 +1,8 @@
-use crate::price::Price;
 use crate::rate::Rate;
-use crate::router::{Channel, PriceFeedId};
 use crate::symbol_state::SymbolState;
 use crate::time::TimestampUs;
+use crate::PriceFeedId;
+use crate::{api::Channel, price::Price};
 use serde::{Deserialize, Serialize};
 use std::time::Duration;
